Use the Intel-provided CMake config scripts
Since 2021, Intel provides a CMake package config file as part of the
oneAPI installs, so lets use that to make sure we have the right
versions/libraries.

Unfortunently, that CMake script isn't added to the path, so we have to
tell CMake where to look using the MKLROOT environment variable set by
the oneAPI setup scripts.